[iOS DnD] [WK2] WKWebView does not handle drag sessions that exit and re-enter https://bugs.webkit.org/show_bug.cgi?id=173856 <rdar://problem/32957100>
Reviewed by Ryosuke Niwa.

Upon exiting a WKContentView, the dropSession tracked by _dataInteractionState is not cleared out. If the drag
session re-enters the WKWebView, a _different_ UIDropSession is sent to the drop delegate, causing us to bail
from handling the session altogether in -dropInteraction:canHandleSession:. To fix this, we need to clear out
the dropSession when the drag session exits the view.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView dropInteraction:sessionDidExit:]):


Identifier: 190711@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218834 268f45cc-cd09-0410-ab3c-d52691b4dbfc