Web Inspector: Initial search is sometimes performed twice, produces duplicate results https://bugs.webkit.org/show_bug.cgi?id=174047
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2017-06-30
Reviewed by Devin Rousso.

* UserInterface/Views/SearchTabContentView.js:
(WebInspector.SearchTabContentView.prototype.performSearch):
The initial layout of the tab content view was setting state to force a search
which may happen after a delay, however the global search field was causing
its own search to happen. We can clear the flag when a search is being performed
to prevent duplicating the search.

Identifier: 190876@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219021 268f45cc-cd09-0410-ab3c-d52691b4dbfc