AVPlayer can continue to be active after released by MediaPlayerPrivateAVFoundationObjC. https://bugs.webkit.org/show_bug.cgi?id=174264
Reviewed by Eric Carlson.

If the AVPlayer is retained (by an autorelease pool, or internally by other objects in
AVFoundation), releasing the AVPlayer is not enough to cancel loading or playback. So before
releasing the AVPlayer, make sure to disassociate the current AVPlayerItem, which should
cancel all activity in the AVPlayer.

* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
(WebCore::MediaPlayerPrivateAVFoundationObjC::cancelLoad):


Identifier: 191109@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219266 268f45cc-cd09-0410-ab3c-d52691b4dbfc