REGRESSION(r210226): Keyboard-focused element not preserved when navigating back through page cache, causing multiple elements to have focus https://bugs.webkit.org/show_bug.cgi?id=174302 <rdar://problem/33204273>
Reviewed by Antti Koivisto.

Source/WebCore:

Don't clear the active/hovered/focused elements when destroying the render tree,
since we might need to reconstruct it later, and would like to remember which
elements those were.

Only the focused state actually stuck when going in and out of the page cache,
but this patch removes all the element pointer clearing for consistency.

Test: fast/history/page-cache-element-state-focused.html

* dom/Document.cpp:
(WebCore::Document::destroyRenderTree):

LayoutTests:

* fast/history/page-cache-element-state-focused-expected.html: Added.
* fast/history/page-cache-element-state-focused.html: Added.


Identifier: 191139@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219296 268f45cc-cd09-0410-ab3c-d52691b4dbfc