[GTK] Spin buttons on input type number appear over the value itself for small widths https://bugs.webkit.org/show_bug.cgi?id=173572
Reviewed by Carlos Garcia Campos.

Source/WebCore:

When drawing the spin buttons, override the width of the input
element to increment it with the width of the spin button.
This ensures that we don't end up covering the input values with
the spin buttons.

Do this also for user controlled styles, because most web authors
won't test how their site renders on WebKitGTK+, and they will
assume spin buttons in the order of 13 pixels wide (that is what
most browsers use), but the GTK+ spin button is much wider (66 pixels).

Test: platform/gtk/fast/forms/number/number-size-spinbutton-nocover.html

* rendering/RenderTheme.cpp:
(WebCore::RenderTheme::adjustStyle):
* rendering/RenderThemeGtk.cpp:
(WebCore::RenderThemeGtk::adjustTextFieldStyle): Call the theme's adjustTextFieldStyle() also for user controlled styles.
(WebCore::RenderThemeGtk::adjustInnerSpinButtonStyle):

LayoutTests:

* fast/forms/number/number-size-expected.txt:
* fast/forms/number/number-size.html: Obtain the size of the spinbutton by inspecting the shadow dom (is more reliable)
  and make the test for user controlled style special for the case of GTK+.
* platform/gtk/fast/forms/number/number-size-spinbutton-nocover-expected.png: Added.
* platform/gtk/fast/forms/number/number-size-spinbutton-nocover-expected.txt: Added.
* platform/gtk/fast/forms/number/number-size-spinbutton-nocover.html: Added.

Identifier: 191175@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219332 268f45cc-cd09-0410-ab3c-d52691b4dbfc