Web Inspector: Node Sidebar - Improve wording for Potential Custom Element lacking a Custom definition https://bugs.webkit.org/show_bug.cgi?id=174346
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2017-07-11
Reviewed by Matt Baker.

* Localizations/en.lproj/localizedStrings.js:
* UserInterface/Views/DOMNodeDetailsSidebarPanel.js:
(WebInspector.DOMNodeDetailsSidebarPanel.prototype._customElementState):
The text "Waiting to be upgraded" is poor. Any non-standard element with dashes
in the name is considered custom, but does not need to be upgraded. Lets use
different wording to indicate that it is detected as custom, and lacks a
specific Custom Element definition.

Identifier: 191176@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219333 268f45cc-cd09-0410-ab3c-d52691b4dbfc