ImageDecoder: Gifs with infinite animation only play once very often https://bugs.webkit.org/show_bug.cgi?id=173403
Reviewed by Michael Catanzaro.

It doesn't always happen, it's easier to reproduce when loading big files from the network, but it also depends
on every file. The problem is that ImageFrameCache is caching the repetition count value always when the size is
already available. In the case of gif files, the loop count value can be at any point of the image stream, so
having the size available doesn't mean we also have the loop count. So, if the value is queried before it's
available, the default value is cached (repeat once) and then always used. We should clear the cached value when
new data is added to the decoder, like we do with other cached values that can change when more data is decoded.

* platform/graphics/ImageFrameCache.cpp:
(WebCore::ImageFrameCache::clearMetadata): Clear m_repetitionCount.

Identifier: 191230@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219389 268f45cc-cd09-0410-ab3c-d52691b4dbfc