Paginated mode: Infinite recursion in RenderTable::layout https://bugs.webkit.org/show_bug.cgi?id=174413
Reviewed by Simon Fraser.

This patch is a workaround for avoiding infinite recursion when the table layout does not stabilize.
Apparently we leak some context (computed padding in this case) from the current to the subsequent layout.
The subsequent layouts always end up producing different line heights for some of the cells in the <thead>.
In paginated mode, when the section moves (<thead>, <tbody> etc) we call layout again recursively.
This could lead to infinite recursion for unstable table layout.

Unable to come up with a reduction yet.

* rendering/RenderTable.cpp:
(WebCore::RenderTable::layout):
* rendering/RenderTable.h:


Identifier: 191235@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219394 268f45cc-cd09-0410-ab3c-d52691b4dbfc