Import test verifying combination of allow-top-navigation and allow-top-navigation-by-user-activation https://bugs.webkit.org/show_bug.cgi?id=174351
Patch by Frederic Wang <fwang@igalia.com> on 2017-07-12
Reviewed by Youenn Fablet.

This imports a test to verify that the combination of the allow-top-navigation and
allow-top-navigation-by-user-activation flags is just treated as allow-top-navigation.
For now the expectation contains a parsing error message which will be removed when bug
171327 is fixed.

* web-platform-tests/html/semantics/embedded-content/the-iframe-element/w3c-import.log:
Add the new test.
* web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_allow_top_navigation-3.html: Added.
This is imported from the W3C repository.
* web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_allow_top_navigation-3-expected.txt: Added.
The test passes, even if for now the 'allow-top-navigation-by-user-activation' triggers a
parsing error message in the console.

Identifier: 191236@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219395 268f45cc-cd09-0410-ab3c-d52691b4dbfc